refactor: format unit test code style
