+++ sdk-component/src/test/java/group/rxcloud/capa/component/configstore/StoreConfigTest.java
@@ -46,5 +46,4 @@ public class StoreConfigTest {
 
         Assertions.assertEquals("30000", storeConfig.getTimeOut());
     }
-
 }
