+++ sdk-component/src/test/java/group/rxcloud/capa/component/configstore/TestCapaConfigStore.java
@@ -42,7 +42,7 @@ public class TestCapaConfigStore extends CapaConfigStore {
 
     @Override
     protected void doInit(StoreConfig storeConfig) {
-        //no extra work for init
+        // no extra work for init
     }
 
     @Override
@@ -100,6 +100,6 @@ public class TestCapaConfigStore extends CapaConfigStore {
 
     @Override
     public void close() throws Exception {
-        //no need to close
+        // no need to close
     }
 }
