+++ sdk-spi/src/test/java/group/rxcloud/capa/spi/http/CapaSerializeHttpSpiTest.java
@@ -298,5 +298,4 @@ public class CapaSerializeHttpSpiTest {
                     Runnable::run);
         }
     }
-
 }
