+++ sdk-spi/src/test/resources/capa-component-rpc-common.properties
@@ -1,0 +1,1 @@
+group.rxcloud.capa.spi.http.config.CapaSpiOptionsLoader=group.rxcloud.capa.spi.http.config.TestSpiOptionsLoader
