--- sdk-spi/src/test/resources/capa-component-rpc.properties
@@ -1,1 +1,0 @@
-group.rxcloud.capa.spi.http.config.CapaSpiOptionsLoader=group.rxcloud.capa.spi.config.TestSpiOptionsLoader
