+++ sdk/src/test/java/group/rxcloud/capa/configuration/CapaConfigurationClientStoreTest.java
@@ -33,9 +33,9 @@ import org.mockito.Mockito;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.util.List;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
