+++ sdk/src/test/java/group/rxcloud/capa/rpc/CapaRpcClientBuilderTest.java
@@ -226,5 +226,4 @@ public class CapaRpcClientBuilderTest {
             };
         }
     }
-
 }
