+++ sdk/src/test/java/group/rxcloud/capa/rpc/CapaRpcClientHttpTest.java
@@ -41,5 +41,4 @@ public class CapaRpcClientHttpTest {
 
         Assertions.assertNotNull(client);
     }
-
 }
